Add subclass for non-event, non-base cards
